Added 64 bit signed integer support to glow::Value and GlowParameterBase. Added additional overloads for unsigned int and unsigned long, which is then encoded as 64 bit integer value.

+++ libember/Headers/ember/dom/impl/AsyncBerReader.ipp
@@ -138,7 +138,7 @@ namespace libember { namespace dom
 
                     case ber::Type::Integer:
                         if (m_length > 4)
-                            return new dom::VariantLeaf(tag, decode<long>());
+                            return new dom::VariantLeaf(tag, decode<long long>());
                         else
                             return new dom::VariantLeaf(tag, decode<int>());
